+++ src/components/Table/MoveValidatorTable.js
@@ -21,7 +21,7 @@ const useStyles = makeStyles({
   table: {
     marginTop: 10,
     marginBottom: 10,
-    maxHeight: 300,
+    maxHeight: 250,
     display: 'flex',
     flexDirection: 'column-reverse'
   },
