+++ src/constants/table/chessOpeningAnalysisTableActionTypes.js
